+++ src/core/overlay/overlay.ts
@@ -88,5 +88,6 @@ export const OVERLAY_PROVIDERS = [
   ViewportRuler,
   OverlayPositionBuilder,
   Overlay,
+  OverlayContainer,
 ];
