+++ src/demo-app/system-config.ts
@@ -3,6 +3,7 @@
  **********************************************************************************************/
 
 const components = [
+  'all',
   'button',
   'card',
   'checkbox',
